Set which encoding your test JVM will start with
This patch adds a setting to the `java.test.config` object.

```jsonc
{
  "java.test.config": {
    "encoding": "ISO-8859-1",
  },
}
```
The aim here is to be able to seamlessly test code which cares
about the launch encoding of the JVM. Yes, such code is not good.
But sometimes you don't control your dependencies enough to fix
this.

Since the debug launcher already cares about the `encoding` member of the
config, the only thing required is to add the member, which is missing
from the launch configs generated by the test plugin.

```typescript
// In `configurationProvider.ts:` (vscode-java-debug)
// VS Code internal console uses UTF-8 to display output by default.
if (config.console === "internalConsole" && !config.encoding) {
    config.encoding = "UTF-8";
}
```

Fixes #1641